hw_1_for_students/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
duckduckgo_search==6.2.4
eval_type_backport==0.2.0
executing==2.1.0
faiss-cpu==1.8.0.post1
fastapi==0.114.1
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
gigachain==0.2.6
gigachain-cli==0.0.25
gigachain-community==0.2.6
gigachain-core==0.2.38.post1
gigachain-text-splitters==0.2.4
gigachat==0.1.35
gigaserve==0.2.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-httplib2==0.2.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.8
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-community==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langfuse==2.48.0
langgraph==0.2.19
langgraph-checkpoint==1.0.12
langsmith==0.1.120
libcst==1.4.0
llama-index==0.9.40
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.45.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.5.0
primp==0.6.1
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pyfiglet==1.0.2
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.0.1
pyproject-toml==0.0.10
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-pptx==0.6.23
python-pptx-interface==0.0.12
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.23.2
wcwidth==0.2.13
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.11.1
